api/requirements.txt
fastapi
fastapi-utils
websockets
uvicorn

bcrypt
pyjwt

redis
aioprometheus